Made it theoretically easier for a compiler to detect 64/32bit operations and optimize accordingly

+++ Library/libc/Sources/stdio/fread.c
@@ -99,7 +99,7 @@ size_t fread(void * _Nonnull _Restrict buffer, size_t size, size_t count, FILE *
 
 
     char* dst = buffer;
-    uint64_t nBytesToRead = (uint64_t)size * (uint64_t)count;
+    uint64_t nBytesToRead = (uint64_t)size * count;
     uint64_t nBytesRead = 0;
     ssize_t r;
 
@@ -116,7 +116,7 @@ size_t fread(void * _Nonnull _Restrict buffer, size_t size, size_t count, FILE *
 
 
     if (nBytesRead > 0ull) {
-        return nBytesRead / (uint64_t)size;
+        return nBytesRead / size;
     }
     else if (r == 0) {
         s->flags.hasEof = 1;
